src/main/java/com/albert/concurrent/ConcurrentService.java
package com.albert.concurrent;

import lombok.Getter;
import lombok.Setter;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;

import java.util.concurrent.ExecutorService;
import java.util.concurrent.Executors;
import java.util.concurrent.Future;

@Slf4j
@Setter
@Getter
public class ConcurrentService {

    private ExecutorService executor
            = Executors.newSingleThreadExecutor();

    public Future<Integer> calculate(Integer input) {
        return executor.submit(() -> {
            log.debug("start calc");
            Thread.sleep(1000);
            log.debug("finished calc");
            return input * input;
        });
    }
}
